+++ packages/spacecat-shared-athena-client/src/traffic-analysis/traffic-data-base-response.js
@@ -33,6 +33,7 @@ export const TrafficDataResponseDto = {
    * }} JSON object.
    */
   toJSON: (data) => ({
+    week: data.week,
     type: data.trf_type,
     channel: data.trf_channel,
     platform: data.trf_platform,