+++ packages/spacecat-shared-data-access/src/models/page-citability/page-citability.model.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Adobe. All rights reserved.
+ * Copyright 2025 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,10 +9,19 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-// eslint-disable-next-line no-console
-console.log('Forcing HTTP/1.1 for Adobe Fetch');
-process.env.HELIX_FETCH_FORCE_HTTP1 = 'true';
-process.env.AWS_ACCESS_KEY_ID = 'fake-key-id';
-process.env.AWS_SECRET_ACCESS_KEY = 'fake-secret';
-// Silence AWS X-Ray context warnings in tests
-process.env.AWS_XRAY_CONTEXT_MISSING = 'IGNORE_ERROR';
+
+import BaseModel from '../base/base.model.js';
+
+/**
+ * PageCitability - Represents a page's citability metrics within a site.
+ *
+ * @class PageCitability
+ * @extends BaseModel
+ */
+class PageCitability extends BaseModel {
+  static DEFAULT_UPDATED_BY = 'spacecat';
+
+  // add any custom methods or overrides here
+}
+
+export default PageCitability;